Add Logger MiddleWare to Push Logs to AWS server
